requirements.txt
aiohappyeyeballs==2.4.4
aiohttp==3.11.11
aiosignal==1.3.2
altair==5.5.0
annotated-types==0.7.0
anyio==4.7.0
attrs==24.3.0
Automat==24.8.1
beautifulsoup4==4.12.3
blinker==1.9.0
cachetools==5.5.0
certifi==2024.12.14
cffi==1.17.1
charset-normalizer==3.4.1
click==8.1.8
constantly==23.10.4
cryptography==44.0.0
cssselect==1.2.0
dataclasses-json==0.6.7
defusedxml==0.7.1
Deprecated==1.2.15
dirtyjson==1.0.8
distro==1.9.0
faiss-cpu==1.9.0.post1
filelock==3.16.1
filetype==1.2.0
frozenlist==1.5.0
fsspec==2024.12.0
gitdb==4.0.11
GitPython==3.1.43
greenlet==3.1.1
groq==0.13.1
h11==0.14.0
httpcore==1.0.7
httpx==0.28.1
httpx-sse==0.4.0
huggingface-hub==0.27.0
hyperlink==21.0.0
idna==3.10
incremental==24.7.2
itemadapter==0.10.0
itemloaders==1.3.2
Jinja2==3.1.5
jiter==0.8.2
jmespath==1.0.1
joblib==1.4.2
jsonpatch==1.33
jsonpointer==3.0.0
jsonschema==4.23.0
jsonschema-specifications==2024.10.1
langchain==0.3.13
langchain-community==0.3.13
langchain-core==0.3.28
langchain-groq==0.2.2
langchain-text-splitters==0.3.4
langsmith==0.2.6
llama-cloud==0.1.7
llama-index==0.12.8
llama-index-agent-openai==0.4.1
llama-index-cli==0.4.0
llama-index-core==0.12.8
llama-index-embeddings-openai==0.3.1
llama-index-indices-managed-llama-cloud==0.6.3
llama-index-llms-openai==0.3.12
llama-index-multi-modal-llms-openai==0.4.1
llama-index-program-openai==0.3.1
llama-index-question-gen-openai==0.3.0
llama-index-readers-file==0.4.1
llama-index-readers-llama-parse==0.4.0
llama-parse==0.5.19
lxml==5.3.0
markdown-it-py==3.0.0
MarkupSafe==3.0.2
marshmallow==3.23.2
mdurl==0.1.2
mpmath==1.3.0
multidict==6.1.0
mypy-extensions==1.0.0
narwhals==1.19.1
nest-asyncio==1.6.0
networkx==3.4.2
nltk==3.9.1
numpy==2.2.1
openai==1.58.1
orjson==3.10.12
outcome==1.3.0.post0
packaging==24.2
pandas==2.2.3
parsel==1.9.1
pillow==11.0.0
playwright==1.49.1
propcache==0.2.1
Protego==0.3.1
protobuf==5.29.2
pyarrow==18.1.0
pyasn1==0.6.1
pyasn1_modules==0.4.1
pycparser==2.22
pydantic==2.10.4
pydantic-settings==2.7.0
pydantic_core==2.27.2
pydeck==0.9.1
PyDispatcher==2.0.7
pyee==12.0.0
Pygments==2.18.0
pyOpenSSL==24.3.0
pypdf==5.1.0
PySocks==1.7.1
python-dateutil==2.9.0.post0
python-dotenv==1.0.1
pytz==2024.2
PyYAML==6.0.2
queuelib==1.7.0
referencing==0.35.1
regex==2024.11.6
requests==2.32.3
requests-file==2.1.0
requests-toolbelt==1.0.0
rich==13.9.4
rpds-py==0.22.3
safetensors==0.4.5
scikit-learn==1.6.0
scipy==1.14.1
Scrapy==2.12.0
selenium==4.27.1
sentence-transformers==3.3.1
service-identity==24.2.0
setuptools==75.6.0
six==1.17.0
smmap==5.0.1
sniffio==1.3.1
sortedcontainers==2.4.0
soupsieve==2.6
SQLAlchemy==2.0.36
streamlit==1.41.1
striprtf==0.0.26
sympy==1.13.1
tenacity==9.0.0
threadpoolctl==3.5.0
tiktoken==0.8.0
tldextract==5.1.3
tokenizers==0.21.0
toml==0.10.2
torch==2.5.1
tornado==6.4.2
tqdm==4.67.1
transformers==4.47.1
trio==0.28.0
trio-websocket==0.11.1
Twisted==24.11.0
typing-inspect==0.9.0
typing_extensions==4.12.2
tzdata==2024.2
urllib3==2.3.0
w3lib==2.2.1
watchdog==6.0.0
websocket-client==1.8.0
wrapt==1.17.0
wsproto==1.2.0
yarl==1.18.3
zope.interface==7.2